refactor(mobile): move contents of Home into app/index.tsx
